fixed size_t issue (its in SPI.h)
Also added vscode to gitignore and fixed comment

+++ bluetooth/nrf52/ChirpBluetoothRxTx.h
@@ -1,7 +1,6 @@
-
 class ChirpBluetoothRxTx
 {
     public:
-        virtual void send(char[], int) = 0;
+        virtual void send(char[], size_t) = 0;
         virtual char* receive() = 0;
 };
